sustainable_tourism_juneau/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
platypus-opt>=1.0.4
scikit-learn>=1.0.0
